backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-generativeai==0.4.0
pydantic==2.6.1
pydantic-settings==2.1.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
python-multipart==0.0.9
aiofiles==23.2.1
# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.26.0
authlib==1.3.0
itsdangerous==2.1.2
email-validator==2.1.0.post1